Series part should be also 0 or decimal

+++ database/migrations/2025_11_30_211219_create_books_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('title');
             $table->string('author');
             $table->foreignId('series_id')->nullable()->constrained();
-            $table->unsignedTinyInteger('part_number')->nullable();
+            $table->double('part_number')->nullable();
             $table->string('publisher');
             $table->unsignedInteger('year');
             $table->foreignId('country_id')->constrained();
